Refactored all keypress login into vminput. This VMInputManager is hosted by the VMCanvas and dispatches Actions to the inner VimMapper struct through Commands. This will enable a VMCanvas to host multiple sheets at some point.
VMInputManager has logic for different key modes to segregate similar
keybinds' functionality. It currently handles mark, jump, and an
unfinished search mode, as well as a general sheet mode. It also has
single-key and string-based keybind support. String keybinds have
captuer groups which define modifiers for custom behavior using a
3-dimensional HashMap.

Added a node move mode triggered by the "`" key. hjkl and HJKL move the
active node when this mode is active.

Added the "gg" keybind to jump to the active node. This is mostly a test
for the string keybind functionality and may change.